--- Schema/Queries/UserQuery.js
@@ -1,12 +1,0 @@
-const { GraphQLList } = require("graphql");
-const { Users } = require("../../models");
-const UserType = require("../TypeDefs/UserType");
-module.exports = {
-  getAllUsers: {
-    type: new GraphQLList(UserType),
-    resolve(parent, args) {
-      let data = Users.findAll();
-      return data;
-    },
-  },
-};
