+++ config/environments/development.json
@@ -1,5 +1,5 @@
 {
-  "port": 3000,
+  "port": 4000,
   "graphql": "/graphql",
   "db": {
     "host": "127.0.0.1",
